requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.9.2
numpy==2.1.2
pandas==2.2.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.1.1
transformers==4.45.2
torch==2.4.1
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==2.1.3
